// ui/src/api/generated/model/deviceKpiResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * Whirlpool Inspection Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.0.3
 */

export interface DeviceKpiResponse {
  total: number;
  active: number;
  deleted: number;
  locked: number;
  unlocked: number;
  desktop: number;
  mobile: number;
}