// ui/src/api/generated/model/deviceListResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * APIs for Whirlpool Inspection Tool API developed by Scopt Analytics for Whirlpool Inspection Tool API platform.
 * OpenAPI spec version: 1.0.3
 */
import type { DeviceResponse } from './deviceResponse';

export interface DeviceListResponse {
  data: DeviceResponse[];
  total: number;
  page: number;
  per_page: number;
  total_pages: number;
}