// ui/src/api/generated/model/deviceResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * APIs for Whirlpool Inspection Tool API developed by Scopt Analytics for Whirlpool Inspection Tool API platform.
 * OpenAPI spec version: 1.0.3
 */

export interface DeviceResponse {
  id: number;
  uuid: string;
  user_id: number;
  user_name: string;
  imei: string;
  device_type: string;
  device_fingerprint: string;
  device_info: unknown;
  ip_address: string | null;
  proxy_ip_address: string | null;
  current_lat: number | null;
  current_lng: number | null;
  is_locked: boolean;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}